requirements.txt
click==8.1.7
pillow==10.3.0
setuptools==69.5.1
tqdm==4.66.2
